// src/app/backend/model/chatMessageDto.ts
/**
 * OutOut API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ChatMessageDataDto } from './chatMessageDataDto';
import { ChatUserDto } from './chatUserDto';


export interface ChatMessageDto { 
    id?: number;
    chatId?: number;
    content?: string | null;
    user?: ChatUserDto;
    createdOn?: string;
    data?: ChatMessageDataDto;
}