// src/app/backend/model/userIdentityImageResponseDto.ts
/**
 * OutOut API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserIdentityImageType } from './userIdentityImageType';


export interface UserIdentityImageResponseDto { 
    id?: number;
    imageUrl?: string | null;
    type?: UserIdentityImageType;
    name?: string | null;
}